local/instrument/metrics/requirements_embedding.txt
fastapi==0.100.0
torch==2.5.1
transformers==4.46.3
uvicorn==0.29.0
pillow==10.4.0
python-multipart==0.0.9
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.17.1